refactor: 💡 change notes to contents
