+++ src/features/viewer/components/viewer-body.tsx
@@ -4,7 +4,7 @@ type Props = {
 	children: ReactNode;
 };
 
-export function ContentBody({ children }: Props) {
+export function ViewerBody({ children }: Props) {
 	return (
 		<div className="prose prose-sm mx-auto p-2 sm:prose-base">{children}</div>
 	);
